+++ Changes.md
@@ -6,6 +6,7 @@
 - Adjust "Intensity" and "Shift" live from Remote for supported presets (eg. Milkwaves shader presets)
 - Set default audio device using Ctrl+D (eg. after disconnecting Bluetooth headphones)
 - New message parameter "fadeout" (also for custom messages)
+- Draw a background box for text messages with custom transparency and color using box_* message parameters
 - New shortcuts: Ctrl+B for toggling button panel, Ctrl+O for opening Visualizer window
 - Force soft preset transition type using "Mixtype" in settings.ini
 - Miscellaneous stability improvements